Error is now within acceptable level. Need to optimize code for run time.

+++ src/particle_filter.h
@@ -17,7 +17,7 @@ struct Particle {
 	double x;
 	double y;
 	double theta;
-	long double weight;
+	double weight;
 	std::vector<int> associations;
 	std::vector<double> sense_x;
 	std::vector<double> sense_y;
